Drop 0.18 support and fix one test
The currenly released version can continue to receive bugfixes if needed.
Not supporting both versions at the same time simplifies the code base
greatly.

This test worked before because the 0.18 version of the code silently
ignored missing exact-deps file. The newer code fails when there's no
matching package for a dependency in ELM_HOME. Fixes by changing the
test to assert the tasks for the project itself *and* dependencies have
been generated.