Make the first move to drop LD_PRELOAD-way of disguising an application as a package
In Elm 0.19, applications still support multiple source directories while packages
do not. In order to generate docs, we need to disguise as a package. The planned
strategy is to copy all source directories to a single src directory inside a
temporary directory. If we're copying files anyways, we can simplify the architecture
and create an elm.json file which is read as normally as it can be without using
an LD_PRELOAD library.