Make fake data configurable and require fake license
Although the license will not be included in the output,
it will be associated with the project during the build,
which is unavoidable due to the Elm compiler enforcing
a valid license when compiling a package project.

The user should be required to explicitly choose a license
as a sign of consent to the above.